develop

+++ backend/src/main/java/com/Adapter/icare/Utils/ExtensionUtils.java
@@ -41,7 +41,7 @@ public class ExtensionUtils {
         return null;
     }
 
-    public static Integer getNestedExtensionValueDecimal(DomainResource resource, String parentUrl, String childUrl) {
+    public static Float getNestedExtensionValueDecimal(DomainResource resource, String parentUrl, String childUrl) {
         if (resource != null && resource.hasExtension()) {
             for (Extension parentExtension : resource.getExtension()) {
                 if (parentExtension.getUrl().equals(parentUrl) && parentExtension.hasExtension()) {
@@ -49,7 +49,7 @@ public class ExtensionUtils {
                         if (childExtension.getUrl().equals(childUrl)
                                 && childExtension.hasValue()
                                 && childExtension.getValue() instanceof DecimalType) {
-                            return ((DecimalType) childExtension.getValue()).getValue().intValue();
+                            return ((DecimalType) childExtension.getValue()).getValue().floatValue();
                         }
                     }
                 }
